FIX: 레포지토리 ID 자료형 변경 및 더미데이터 추가

+++ Server/src/main/java/com/threeracha/gaewoonhae/db/repository/PointHistoryRepository.java
@@ -9,7 +9,7 @@ import java.util.Date;
 import java.util.Optional;
 
 @Repository
-public interface PointHistoryRepository extends JpaRepository<PointHistory, Integer> {
+public interface PointHistoryRepository extends JpaRepository<PointHistory, Long> {
     Optional<PointHistory> findByUserUserIdAndChangeTime(Long userId, Date changeTime);
     Optional<PointHistory> findByUserUserIdAndPointChange(Long userId, int pointChange);
